requirements-dev.txt
anthropic
click
google-genai
instructor
openai
pydantic
PyYAML
tenacity
typer
pytest
pre-commit
ruff
